// src/lib/apis/tmdb/tmdb.generated.d.ts
export interface paths {
  "/3/search/movie": {
    /**
     * Movie 
     * @description Search for movies by their original, translated and alternative titles.
     */
    get: operations["search-movie"];
  };
  "/3/discover/movie": {
    /**
     * Movie 
     * @description Find movies using over 30 filters and sort options.
     */
    get: operations["discover-movie"];
  };
  "/3/movie/{movie_id}": {
    /**
     * Details 
     * @description Get the top level details of a movie by ID.
     */
    get: operations["movie-details"];
  };
// ...
  "/3/find/{external_id}": {
    /**
     * Find By ID 
     * @description Find data by external ID's.
     */
    get: operations["find-by-id"];
  };
// ...
}

export type webhooks = Record<string, never>;

export interface components {
  schemas: never;
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}

export type external = Record<string, never>;

export interface operations {
// ...
  "movie-details": {
    parameters: {
      query?: {
        /** @description comma separated list of endpoints within this namespace, 20 items max */
        append_to_response?: string;
        language?: string;
      };
      path: {
        movie_id: number;
      };
    };
    responses: {
      /** @description 200 */
      200: {
        content: {
          "application/json": {
            /**
             * @default true 
             * @example false
             */
            adult?: boolean;
            /** @example /hZkgoQYus5vegHoetLkCJzb17zJ.jpg */
            backdrop_path?: string;
            belongs_to_collection?: unknown;
            /**
             * @default 0 
             * @example 63000000
             */
            budget?: number;
            genres?: ({
                /**
                 * @default 0 
                 * @example 18
                 */
                id?: number;
                /** @example Drama */
                name?: string;
              })[];
            /** @example http://www.foxmovies.com/movies/fight-club */
            homepage?: string;
            /**
             * @default 0 
             * @example 550
             */
            id?: number;
            /** @example tt0137523 */
            imdb_id?: string;
            /** @example en */
            original_language?: string;
            /** @example Fight Club */
            original_title?: string;
            /** @example A ticking-time-bomb insomniac and a slippery soap salesman channel primal male aggression into a shocking new form of therapy. Their concept catches on, with underground "fight clubs" forming in every town, until an eccentric gets in the way and ignites an out-of-control spiral toward oblivion. */
            overview?: string;
            /**
             * @default 0 
             * @example 61.416
             */
            popularity?: number;
            /** @example /pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg */
            poster_path?: string;
            production_companies?: ({
                /**
                 * @default 0 
                 * @example 508
                 */
                id?: number;
                /** @example /7cxRWzi4LsVm4Utfpr1hfARNurT.png */
                logo_path?: string;
                /** @example Regency Enterprises */
                name?: string;
                /** @example US */
                origin_country?: string;
              })[];
            production_countries?: ({
                /** @example US */
                iso_3166_1?: string;
                /** @example United States of America */
                name?: string;
              })[];
            /** @example 1999-10-15 */
            release_date?: string;
            /**
             * @default 0 
             * @example 100853753
             */
            revenue?: number;
            /**
             * @default 0 
             * @example 139
             */
            runtime?: number;
            spoken_languages?: ({
                /** @example English */
                english_name?: string;
                /** @example en */
                iso_639_1?: string;
                /** @example English */
                name?: string;
              })[];
            /** @example Released */
            status?: string;
            /** @example Mischief. Mayhem. Soap. */
            tagline?: string;
            /** @example Fight Club */
            title?: string;
            /**
             * @default true 
             * @example false
             */
            video?: boolean;
            /**
             * @default 0 
             * @example 8.433
             */
            vote_average?: number;
            /**
             * @default 0 
             * @example 26280
             */
            vote_count?: number;
          };
        };
      };
    };
  };
// ...
  "movie-credits": {
    parameters: {
      query?: {
        language?: string;
      };
      path: {
        movie_id: number;
      };
    };
    responses: {
      /** @description 200 */
      200: {
        content: {
          "application/json": {
            /**
             * @default 0 
             * @example 550
             */
            id?: number;
            cast?: ({
                /**
                 * @default true 
                 * @example false
                 */
                adult?: boolean;
                /**
                 * @default 0 
                 * @example 2
                 */
                gender?: number;
                /**
                 * @default 0 
                 * @example 819
                 */
                id?: number;
                /** @example Acting */
                known_for_department?: string;
                /** @example Edward Norton */
                name?: string;
                /** @example Edward Norton */
                original_name?: string;
                /**
                 * @default 0 
                 * @example 26.99
                 */
                popularity?: number;
                /** @example /8nytsqL59SFJTVYVrN72k6qkGgJ.jpg */
                profile_path?: string;
                /**
                 * @default 0 
                 * @example 4
                 */
                cast_id?: number;
                /** @example The Narrator */
                character?: string;
                /** @example 52fe4250c3a36847f80149f3 */
                credit_id?: string;
                /**
                 * @default 0 
                 * @example 0
                 */
                order?: number;
              })[];
            crew?: ({
                /**
                 * @default true 
                 * @example false
                 */
                adult?: boolean;
                /**
                 * @default 0 
                 * @example 2
                 */
                gender?: number;
                /**
                 * @default 0 
                 * @example 376
                 */
                id?: number;
                /** @example Production */
                known_for_department?: string;
                /** @example Arnon Milchan */
                name?: string;
                /** @example Arnon Milchan */
                original_name?: string;
                /**
                 * @default 0 
                 * @example 2.931
                 */
                popularity?: number;
                /** @example /b2hBExX4NnczNAnLuTBF4kmNhZm.jpg */
                profile_path?: string;
                /** @example 55731b8192514111610027d7 */
                credit_id?: string;
                /** @example Production */
                department?: string;
                /** @example Executive Producer */
                job?: string;
              })[];
          };
        };
      };
    };
  };
  /** External IDs */
  "movie-external-ids": {
    parameters: {
      path: {
        movie_id: number;
      };
    };
    responses: {
      /** @description 200 */
      200: {
        content: {
          "application/json": {
            /**
             * @default 0 
             * @example 550
             */
            id?: number;
            /** @example tt0137523 */
            imdb_id?: string;
            wikidata_id?: unknown;
            /** @example FightClub */
            facebook_id?: string;
            instagram_id?: unknown;
            twitter_id?: unknown;
          };
        };
      };
    };
  };
// ...
  "movie-videos": {
    parameters: {
      query?: {
        language?: string;
      };
      path: {
        movie_id: number;
      };
    };
    responses: {
      /** @description 200 */
      200: {
        content: {
          "application/json": {
            /**
             * @default 0 
             * @example 550
             */
            id?: number;
            results?: ({
                /** @example en */
                iso_639_1?: string;
                /** @example US */
                iso_3166_1?: string;
                /** @example Fight Club (1999) Trailer - Starring Brad Pitt, Edward Norton, Helena Bonham Carter */
                name?: string;
                /** @example O-b2VfmmbyA */
                key?: string;
                /** @example YouTube */
                site?: string;
                /**
                 * @default 0 
                 * @example 720
                 */
                size?: number;
                /** @example Trailer */
                type?: string;
                /**
                 * @default true 
                 * @example false
                 */
                official?: boolean;
                /** @example 2016-03-05T02:03:14.000Z */
                published_at?: string;
                /** @example 639d5326be6d88007f170f44 */
                id?: string;
              })[];
          };
        };
      };
    };
  };
// ...
  "tv-series-external-ids": {
    parameters: {
      path: {
        series_id: number;
      };
    };
    responses: {
      /** @description 200 */
      200: {
        content: {
          "application/json": {
            /**
             * @default 0 
             * @example 1399
             */
            id?: number;
            /** @example tt0944947 */
            imdb_id?: string;
            /** @example /m/0524b41 */
            freebase_mid?: string;
            /** @example /en/game_of_thrones */
            freebase_id?: string;
            /**
             * @default 0 
             * @example 121361
             */
            tvdb_id?: number;
            /**
             * @default 0 
             * @example 24493
             */
            tvrage_id?: number;
            /** @example Q23572 */
            wikidata_id?: string;
            /** @example GameOfThrones */
            facebook_id?: string;
            /** @example gameofthrones */
            instagram_id?: string;
            /** @example GameOfThrones */
            twitter_id?: string;
          };
        };
      };
    };
  };
// ...
  "tv-series-videos": {
    parameters: {
      query?: {
        /** @description filter the list results by language, supports more than one value by using a comma */
        include_video_language?: string;
        language?: string;
      };
      path: {
        series_id: number;
      };
    };
    responses: {
      /** @description 200 */
      200: {
        content: {
          "application/json": {
            /**
             * @default 0 
             * @example 1399
             */
            id?: number;
            results?: ({
                /** @example en */
                iso_639_1?: string;
                /** @example US */
                iso_3166_1?: string;
                /** @example Inside Game of Thrones: A Story in Camera Work – BTS (HBO) */
                name?: string;
                /** @example y2ZJ3lTaREY */
                key?: string;
                /** @example YouTube */
                site?: string;
                /**
                 * @default 0 
                 * @example 1080
                 */
                size?: number;
                /** @example Behind the Scenes */
                type?: string;
                /**
                 * @default true 
                 * @example true
                 */
                official?: boolean;
                /** @example 2019-03-25T14:00:06.000Z */
                published_at?: string;
                /** @example 5c999b48c3a36863b73b9d42 */
                id?: string;
              })[];
          };
        };
      };
    };
  };
// ...
}